Install (and uninstall) script for icons
The scripts take care of installing the icon on Mac systems. It will
back up the original icon, copy the new icon in and symlink it, touch
the app to refresh it, and restart the Dock. On the next boot of the
app, the icon will be new.

Uninstall does this process in reverse and restores the original icon.